+++ src/combinators/switchMap.test.js
@@ -1,0 +1,88 @@
+import { always, id } from 'fkit'
+
+import Signal from '../Signal'
+import switchMap from './switchMap'
+
+let valueSpy, errorSpy, completeSpy
+
+describe('switchMap', () => {
+  beforeEach(() => {
+    valueSpy = jest.fn()
+    errorSpy = jest.fn()
+    completeSpy = jest.fn()
+  })
+
+  it('applies a function to the signal values', () => {
+    let value
+    const s = new Signal(emit => {
+      value = emit.value
+    })
+    const f = a => Signal.of(a)
+
+    switchMap(f, s).subscribe(valueSpy, errorSpy, completeSpy)
+
+    expect(valueSpy).not.toHaveBeenCalled()
+    value(1)
+    expect(valueSpy).toHaveBeenCalledTimes(1)
+    expect(valueSpy).toHaveBeenLastCalledWith(1)
+    value(2)
+    expect(valueSpy).toHaveBeenCalledTimes(2)
+    expect(valueSpy).toHaveBeenLastCalledWith(2)
+    value(3)
+    expect(valueSpy).toHaveBeenCalledTimes(3)
+    expect(valueSpy).toHaveBeenLastCalledWith(3)
+  })
+
+  it('emits an error when the given signal emits an error', () => {
+    let error
+    const s = new Signal(emit => {
+      error = emit.error
+    })
+
+    switchMap(always(), s).subscribe(valueSpy, errorSpy, completeSpy)
+
+    expect(errorSpy).not.toHaveBeenCalled()
+    error('foo')
+    expect(errorSpy).toHaveBeenCalledTimes(1)
+    expect(errorSpy).toHaveBeenCalledWith('foo')
+  })
+
+  it('completes when the given signal is completed', () => {
+    let completeS, completeT
+    const s = new Signal(emit => {
+      completeS = emit.complete
+    })
+    const t = new Signal(emit => {
+      emit.value(s)
+      completeT = emit.complete
+    })
+
+    switchMap(id, t).subscribe(valueSpy, errorSpy, completeSpy)
+
+    completeS()
+    expect(completeSpy).not.toHaveBeenCalled()
+    completeT()
+    expect(completeSpy).toHaveBeenCalledTimes(1)
+  })
+
+  it('unmounts the given signal when the returned signal is unsubscribed', () => {
+    const unmount = jest.fn()
+    const s = new Signal(() => unmount)
+    const a = switchMap(id, s).subscribe()
+
+    expect(unmount).not.toHaveBeenCalled()
+    a.unsubscribe()
+    expect(unmount).toHaveBeenCalledTimes(1)
+  })
+
+  it('unmounts the child signal when the returned signal is unsubscribed', () => {
+    const unmount = jest.fn()
+    const s = new Signal(() => unmount)
+    const t = Signal.of(s)
+    const a = switchMap(id, t).subscribe()
+
+    expect(unmount).not.toHaveBeenCalled()
+    a.unsubscribe()
+    expect(unmount).toHaveBeenCalledTimes(1)
+  })
+})